requirement.txt
langchain
python-dotenv
langchain_community
streamlit
openai
langchain-openai